Add manifests generation verification script

+++ hack/update-lws-controller-manifests.sh
@@ -17,10 +17,11 @@ set -o nounset
 set -o pipefail
 
 SCRIPT_ROOT="$(realpath "$(dirname "$(readlink -f "$0")")"/..)"
+LWS_BRANCH="${LWS_BRANCH:-main}"
 LWS_ASSETS_DIR="${SCRIPT_ROOT}/bindata/assets/lws-controller-generated"
 LWS_CONTROLLER_DIR="${LWS_CONTROLLER_DIR:-${HOME}/go/src/sigs.k8s.io/lws}"
 
-LWS_RELEASE_TAG="${LWS_RELEASE_TAG:-"upstream/main"}" # "v0.5.0"
+LWS_RELEASE_TAG="${LWS_RELEASE_TAG:-"upstream/$LWS_BRANCH"}" # "v0.5.0"
 LWS_NAMESPACE="${LWS_NAMESPACE:-openshift-lws-operator}"
 
 if [ ! -d "${LWS_CONTROLLER_DIR}" ]; then
@@ -63,7 +64,11 @@ pushd "${LWS_ASSETS_DIR}"
 # we need to modify prometheus rolebinding to use openshift-monitoring namespace
 sed -i 's/namespace: monitoring/namespace: openshift-monitoring/g' rbac.authorization.k8s.io_v1_rolebinding_lws-prometheus-k8s.yaml
 # we supply our own config
-rm ./v1_configmap_lws-manager-config.yaml
+if [ -e "./v1_configmap_lws-manager-config.yaml" ]; then
+  rm ./v1_configmap_lws-manager-config.yaml
+fi
 # we don't need the namespace object
-rm ./v1_namespace_openshift-lws-operator.yaml
+if [ -e "./v1_namespace_openshift-lws-operator.yaml" ]; then
+  rm ./v1_namespace_openshift-lws-operator.yaml
+fi
 popd
